Remove unnecessary stuff, use correct devtools setup
