+++ .travis.yml
@@ -1,30 +1,44 @@
-language: python
+language: c
+sudo: false
+addons:
+  apt:
+    sources:
+    #- ubuntu-toolchain-r-test
+    packages:
+      - ca-certificates # needed for travis cache to work
 
-notifications:
-  email: false
 
-before_install:
-  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - conda config --set always_yes yes
-  - conda update conda
-  - conda install conda-build
-  - conda config --add channels omnia
-  - conda info -a
+cache:
+  directories:
+    - $HOME/.cache/msmpipeline
 
 install:
-  - conda create --name test-environment python=$TRAVIS_PYTHON_VERSION pyemma msmbuilder
-  - source activate test-environment
-  - conda build devtools/conda-recipe
-  - python setup.py install
+  - source devtools/travis-ci/install.sh
+  - export PYTHONUNBUFFERED=true
+  - export CC=gcc
+  - export CXX=g++
+  # This line may be needed to get travis cache to work
+  - echo "cacert=/etc/ssl/certs/ca-certificates.crt" >> ~/.curlrc
 
 script:
-  - quick_test
-  - test-report
+  - conda config --add channels ${ORGNAME}
+  # Add omnia dev channels
+  - conda config --add channels https://conda.anaconda.org/omnia/label/dev
+  - pip install glob
+  - conda build devtools/conda-recipe
+  - source activate _test
+  - msm-pipeline "msmpipeline/test_data/alanine_*.h5" "alanine"
 
 env:
   matrix:
     - python=2.7  CONDA_PY=27
     - python=3.4  CONDA_PY=34
     - python=3.5  CONDA_PY=35
+
+  global:
+    - PACKAGENAME="msmpipeline"
+    - ORGNAME="omnia"
+
+after_success:
+  - echo "after_success"
+  - ./devtools/travis-ci/after_success.sh
