+++ devtools/travis-ci/install.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# Install Miniconda
+MINICONDA=Miniconda2-latest-Linux-x86_64.sh
+MINICONDA_MD5=$(curl -s http://repo.continuum.io/miniconda/ | grep -A3 $MINICONDA | sed -n '4p' | sed -n 's/ *<td>\(.*\)<\/td> */\1/p')
+wget http://repo.continuum.io/miniconda/$MINICONDA
+if [[ $MINICONDA_MD5 != $(md5sum $MINICONDA | cut -d ' ' -f 1) ]]; then
+    echo "Miniconda MD5 mismatch"
+    exit 1
+fi
+bash $MINICONDA -b -p $HOME/miniconda
+
+# Configure miniconda
+export PATH=$HOME/miniconda/bin:$PATH
+conda install --yes conda-build anaconda-client jinja2 pip
