+++ msmpipeline/__init__.py
@@ -1,4 +1,1 @@
-# versioneer
-from ._version import get_versions
-__version__ = get_versions()['version']
-del get_versions
+from .msmpipeline import *
