--- msmpipeline/quick_test.py
@@ -1,5 +1,0 @@
-def test_generate_test_instance():
-    from test_report_generation import generate_test_instance
-    _ = generate_test_instance()
-
-test_generate_test_instance()
