+++ setup.py
@@ -1,8 +1,7 @@
 from setuptools import setup
-import versioneer
 
 setup(name='msmpipeline',
-      version=versioneer.get_version(),
+      version='0.0.1',
       description='Sensible defaults for MSMs',
       url='https://github.com/choderalab/msm-pipeline/',
       author='Sonya Hanson, Steven Albanese, Josh Fass',
@@ -11,9 +10,7 @@ setup(name='msmpipeline',
       packages=['msmpipeline'],
       entry_points={
           'console_scripts': [
-              'msm-pipeline = msmpipeline.pipeline:main',
-              'test-report = msmpipeline.test_report_generation:main',
-              'quick_test = msmpipeline.quick_test:test_generate_test_instance'
+              'msm-pipeline = msmpipeline:main'
           ]
       },
       zip_safe=False)
